requirements.txt
py7zr>=0.20.5
python-dotenv>=1.0.0
